+++ Content/Enemies/Standard/AlphaConstruct.cs
@@ -4,16 +4,16 @@
     {
         public override string Name => ":: Enemies :: Alpha Construct";
 
+        [ConfigField("Should replace Blind Pests on Siphoned Forest?", "", false)]
+        public static float shouldReplace;
+
+        [ConfigField("Should spawn on Commencement?", "", false)]
+        public static float shouldSpawn;
+
         public override void Init()
         {
             base.Init();
         }
-
-        [ConfigField("Should replace Blind Pests on Siphoned Forest?", "", true)]
-        public static bool shouldReplaceBlindPestsOnSiphonedForest;
-
-        [ConfigField("Should spawn on Commencement?", "", false)]
-        public static bool shouldSpawnOnCommencement;
 
         public override void Hooks()
         {
@@ -30,34 +30,6 @@
 
         private void Changes()
         {
-            if (shouldSpawnOnCommencement)
-            {
-                var commencement = Utils.Paths.DirectorCardCategorySelection.dccsMoonMonsters.Load<DirectorCardCategorySelection>();
-                var commencementDLC1 = Utils.Paths.DirectorCardCategorySelection.dccsMoonMonstersDLC1.Load<DirectorCardCategorySelection>();
-
-                var alphaConstructCard = new DirectorCard
-                {
-                    spawnCard = Utils.Paths.CharacterSpawnCard.cscMinorConstruct.Load<CharacterSpawnCard>(),
-                    selectionWeight = 1,
-                    spawnDistance = DirectorCore.MonsterSpawnDistance.Close,
-                    preventOverhead = false
-                };
-
-                var cardHolder = new DirectorAPI.DirectorCardHolder
-                {
-                    Card = alphaConstructCard,
-                    MonsterCategory = DirectorAPI.MonsterCategory.BasicMonsters
-                };
-
-                commencement.AddCard(cardHolder);
-                commencementDLC1.AddCard(cardHolder);
-            }
-
-            if (shouldReplaceBlindPestsOnSiphonedForest)
-            {
-                var siphoned = Utils.Paths.DirectorCardCategorySelection.dccsSnowyForestMonstersDLC1.Load<DirectorCardCategorySelection>();
-                siphoned.categories[2] /* standard monsters */.cards[2] /* blind pest */.spawnCard = Utils.Paths.CharacterSpawnCard.cscMinorConstruct.Load<CharacterSpawnCard>();
-            }
         }
     }
 }