+++ Content/Interactables/ScavengerBag.cs
@@ -10,6 +10,9 @@
         [ConfigField("Lunar Coin Drop Count", "Only applies to Twisted Scavengers.", 10)]
         public static int lunarCoinDropCount;
 
+        [ConfigField("Enable Infinite Scavenger Bags?", "", true)]
+        public static bool infiniteBags;
+
         public override void Init()
         {
             base.Init();
@@ -18,6 +21,20 @@
         public override void Hooks()
         {
             On.EntityStates.ScavBackpack.Opening.OnEnter += Opening_OnEnter;
+            On.EntityStates.ScavMonster.Death.OnEnter += Death_OnEnter;
+        }
+
+        private void Death_OnEnter(On.EntityStates.ScavMonster.Death.orig_OnEnter orig, EntityStates.ScavMonster.Death self)
+        {
+            if (NetworkServer.active)
+            {
+                Stage.instance.scavPackDroppedServer = false;
+            }
+            orig(self);
+            if (NetworkServer.active)
+            {
+                Stage.instance.scavPackDroppedServer = false;
+            }
         }
 
         private void Opening_OnEnter(On.EntityStates.ScavBackpack.Opening.orig_OnEnter orig, EntityStates.ScavBackpack.Opening self)